Adding validations rules for training flights.
This is to comply with rules set by the recipient of the training flights (first case: DSvU DTO / ForeningsAdministration / ElevMappen).
Training Exercises (in code: Training2Lesson) can be marked as valid for solo/dual flight (none of these set implies that this is a briefing-only exercise).
TrainingPrograms may be marked as requiring UnionIDs for student and instructor.

When editing a training flight, validation issues are shown at the top.
If a validation issue is present for a flight, this is indicated in the training flight history page.

The implementation should work also in cases (other DTOs) where these rules do not apply.